requirements.txt
jinja2
python-gnupg
pyyaml
requests
markdown